example/requirements.txt
playwright
pyautogui
opencv-python
pywinauto; platform_system == "Windows"
python-dotenv
loguru
schedule
